Unify error message for calling .new() instead of :new().
Also do not longer treat the receiver as argument #0, but as
argument #1. This makes things easier, because Lua’s luaL_check*
functions generate error messages containing the stack index, which
then causes a gap -- #0 is the receiver, #2 the first argument? Just
assume now that users now enough Lua to grasp the difference between
the . and : syntaxes.