Convert to CMake as the build system and include Lua as a subproject.
Properly including Lua as a subproject should fix away all those
compilation problems due to conflicts with system Lua, etc. once
and for all.

The conversion to CMake was quite easy, with the exception of CEGUI,
which didn’t had a CMake package file by default, but one was easily
findable on the Internet (see smc/cmake/modules/FindCEGUI.cmake for
copyright).

Last but not least this only does compilation now, not yet installation.
This means that due to the lack of graphics, etc. the game won’t run
yet.