Wrote a parser for C++ files that extracts comments about our scripting system.
Maintaining the scripting documentation right where the implementation
is eases the development process and prevents getting the docs out
of sync with the reality more effectively. Originally I just wanted
to use RDoc, but this didn’t work as it’s C parse is tight too closely
to MRI C extensions. And we already had a nice docs layout, so I wanted
to re-use that. And writing parsers is fun anyway :-).

The parser is however very rudimentary. If you don’t stick exactly
to the comment format you will see in the scripting sourcefiles
hopefully soon, it will most likely crash with some obscure errors.