Implement much more performent UID generation on level startup.
Previously, loading larger levels (e.g. the game tutorial level)
resulted in a very long loading time, because each time an UID
had to be generated, all sprites already in the level hat to
be iterated and checked for a free UID. Now we maintain a sorted,
dynamically-growing list of free UIDs instead, so we just need to
pick the first element from that list to find the first free UID.
This speeds up level loading tremendously.