Show backtrace on exceptions.
Sadly, this seems to not work correctly on the mruby side (yet).
Printing a backtrace from the C side simply does not work — you
have to put a begin/rescue somewhere and use the rescue clause to
print the exception. However, this might be fixed in the future and
the backtrace printing code doesn’t hurt (just prints an empty line),
so I’ll just leave it in and it hopefully starts to work with future
versions of mruby.