Change overworld loading to post-initialization again.
The broken design of cOverworld_description enforces this; it
references information that requires a preexisting cOverworld instance,
which a cOverworld instance needs a cOverworld_description instance
to be complete.
The solution for this is to remove the m_path attribute from
cOverworld_description (where it really does not belong) and
add it to cOverworld directly instead. Then we could have a
cOverworld constructor that could create the necessary
cOverworld_description nicely in the way it is intended to be
created (read: proper OOP composition).