Stop infinitely repeating level win music. Fixes #21.
The previous code tried to guess whether the main menu was activated
by looking at what music was playing currently. The new code now looks
if we’re inside a level, which works far more reliably. However, there
seems to be no possibility to check whether we’re in an overworld;
to trigger this bug, you have to enter the menu from within an overworld
and completely let the overworld music finish off. After that, the main
theme will start, but as soon as you leave the menu, the overworld music
will reappear. I doubt this is a large problem, given that the overworld
music themes are quite long and you would therefore have to wait quite
long in the menu for that bug to appear.