Fix giant amount of whitespaces being inserted into the <script> tag.
The level loader treated ALL characters as script tags, including newlines,
because I forgot to initialize m_in_script_tag to false. It defaulted to
true, and so the on_characters() member just added everything to
the script on loading. When saving the level, all these empty lines
would be written out to the level file.