Rename cSprite_Manager::Set_Pos_Z() to cSprite_Manager::Ensure_Different_Z() and explain what it does.
And make it private. This is not meant to be called from outside
the sprite manager. Also fix it to properly work with massive
types instead of general types.

Currently, this causes enemies to be given Z coordinates in the same
line as other sprites, so that e.g. flyons appear before their pipe.