Initial changes in our attempt to upgrade from CEGUI 0.7 to 0.8
It successfully compiles, but I haven't been able to test it because I'm
having trouble getting the build system to properly do staged install. There
are two lines that I commented out with a TODO that I just couldn't figure out
how to port. I'll look more into it later, but this should be a good initial
staging point for further efforts.