Circumvent serious segfault incompatibility between SDL_image and FreeImage.
It appears that the SDL_image library conflicts with newer versions
of FreeImage, culminating in a segmentation fault in libPNG, which
was quite a difficult thing to figure out. FreeImage ships with an
internal libPNG library, which is incompatible with the current
libPNG 1.6.10 and which leads to unpredictable segmentation faults
when SDL_image tries to use libPNG functions. The visible result
was SMC crashing directly on startup when calling SDL’s IMG_Load()
function.

As FreeImage directly includes the faulty libPNG and SDL_image has
a hard dependency on libPNG in a different version, there is an
absolute and unbridgeable incompatibility between recent SDL_image
and FreeImage libraries. The only "solution" to this is to completely
remove the FreeImage library entirely, which was in turn a bit
difficult as CEGUI depends on FreeImage as an image codec. Fortunately,
CEGUI provides other image codecs, and for the meantime I chose
to use the DevIL codec, which works just fine (segfaults are gone now).

As a consequence, FreeImage is no longer a dependency of SMC, but
the DevIL library now is. DevIL is not actively developed currently,
but there seem to be no real other options for CEGUI to work with.