Fix file-not-found exception when clicking “Reload” in the level editor.
It’s not useful to have the reload function re-execute the search
for the level file anyway, as the case that either the game level
file or the user level file suddenly disappears will never happen.
It will always be the very same file that gets loaded, so we can
safely skip the search process and just rely on the already existing
filename.

Also note that the reload function isn’t ideal. See the sourcecode
comments on this.