Fixed crash when changing package.
This is a fix for a crash that occurs when loading a package.  When
loading a package, the current list of overworlds must be cleared and
reloaded from the newly selected package.  This deletes the current
overworld "pActive_Overworld", and resetting the player's save will
attempt to load World 1. If it does not exist then a blank world will
be created to avoid a crash.