Changes to package manager, some level work.
Some minor changes to the package manager are that levels, worlds and
campaigns are now package-only, meaning that they do not use the search
path.  It seems to make sense that a package that provides these items
would have their own "story line" and not depend on a story line of
another package.  Media files still use the search path.  This makes it
possible to create resource packages that contain graphics and music
that can be used by other packages.

The package manager now supports saving levels in the current package.
In addition, a music filename can be set which will use the package
search path.