Scripting can load scripts from packages.
This commit changes the scripting call of SMC.require to take an extra
optional argument.  Without the argument, it loads scripts only from the
game scripting directory.  If the package argument is specified, the
user and game scripting paths associated with that package will be
searched for the script.

This represents the normal usage of only game scripting paths:

    SMC.require "std/all"

This represents using a script from a package.  It can be the same
package that the level is in, or a different package:

    SMC.require "init", "simpletoon/levelset1"