For the *nix version, compile in CEGUI statically.
This means that in the build process we now download and compile
CEGUI specifically for static linking. This is required currently
because the old CEGUI we need is not shipped by any modern Linux
distribution anymore, so requiring users to link it dynamically
would break their other programs that depend on CEGUI 0.8.x.

The windows version still uses MXE’s default CEGUI, as that one
is still the old 0.7.x CEGUI.