Ensure crates don’t hang in the air after they killed an enemy.
Cf. ticket #226. This commit does not solve the problem entirely,
because for some reason a falling crate doesn’t always kill the
enemy below as it should do. It then lies on the enemy and gets
carried around by it. If you kill the enemy (e.g. by fireball),
the crate will hang in the air again.