Reorganize authors.txt file.
It now gives a much better overview of all the copyright holders
plus conact information (the latter is stripped out for the credits
screen in the game). Specialthanks people who have helped the game,
but not contributed code or assets, hence do not hold copyright,
have been extracted into a separate file (which is also read for
the credits screen).