Add updated settings file for small hill
The image was removed as part of #256, but
this caused missing images in older levels.
Now any instances of it in levels will load
the referenced image, small_1, and show as
obsolete (red background) in the editor.
The sprite isn't included in the editor menu,
so new levels won't use it accidentally.