Static enemy now uses Add_Animation
When loading, if there is an "animation" property, it overrides the
"image" property.  When saving, the "animation" property will be created
if the image filename does not end with extension ".png".  If an
animaiton is specified for the static enemy, the animation will be
played while the enemy rotates, moves along the path, etc.