Fix debug message in animation subsystem
The RGBA values are stored as character values (type char) and therefore
are printed as characters instead of integers. Since we actually want
integer values, we need to convert them to type int. This is the nature
of std::cout/cerr message printing.

The file name doesn't seem to be printed correctly either, but I don't
know the cause of that.