Allow image set reference to use search path.
This change allows the path specified in an image set to use the search
path for images.  For example:

An image set locatated at /package1/pixmaps/mine/my.imgset can contain:

    ../blocks/ice/1.png
    ../blocks/ice/2.png

Originally, this would have forced the images used to be:

    /packace1/pixmaps/mine/../blocks/ice/1.png
    /package1/pixmaps/mine/../blocks/ice/2.png

The change stores the relative path of the image set, in this case
mine/my.imgset, so the images used will be:

    mine/../blocks/ice/1.png
    mine/../blocks/ice/2.png

Now, the problem is that if mine/ does not exist in a specific package
search path, then it will not be used.  For instance:

    /basedata/pixmaps/mine/../blocks/ice/1.png

Will fail to load because /basedata/pixmaps/mine/ does not exist. The
solution is to normalize the image path with the data file relative
path:

    mine/ + ../blocks/ice/1.png = blocks/ice/1.png

However, this means that symbolic links will not be followed.  So
it is best not to use symbolic links for any of the data.