Adapt cBaseBox to new savegame mechanism.
This commit removes the check for the item count as the decision
for whether to save the box or not. It complicates the code with
regard to subclasses unnecessary, and the advantage of only saving
a changed box into the savegame is so marginal that it can be neglected
without further impact. Just always save boxes, so subclasses can
just call the parent method and don’t have to expect missing
attributes.