Do not Copy() regular objects during saving.
Copy() doesn’t do what’s expected. It doesn’t cascade upwards properly
and saves things like the m_active attribute. For saving using the
very same object that is in the level is just fine also, as it
doesn’t get modified during saving. const guarantee is intended to
be added in a later commit. See #344 for the ticket regarding the
Copy() problem..