Decouple image manager from cLevel lifetime.
This makes it possible to create cSprite_Actor subclass instances
even when no or another level is active. Which is very useful for
level preperation, because usual level activation order is like
this:

  1. Instanciate cLevel
  2. Create instances of cActor
  3. Add cActor instances into the level
  4. Activate the level instance

Previously, step 2 depended on step 4, which was resolved by
(ab)using the Added_To_Level() callback so that actual initialisation
was deferred. The entire problem is now gone by just having the
image manager available separate from the level.