Protect only the actual movement during warping. Fixes #431.
Before this commit, the warp protection introduced for fixing #358
was only released by a LevelEntry corresponding to a LevelExit,
and that LevelEntry needed to be a movement one (and not a
LevelEntryBeam). This commit moves the entire warping kill
protection solely into LevelExit, such that it only surrounds the
actual movement during use of the LevelExit. This way the section
covered by the warp protection is more predictable as well, given
that TSC uses the same cLevel_Player instance throughout the entire
game.