Extract variable initialisation into separate Init() methods.
Note that in C++, you can’t call virtual functions in a constructor,
because the subclass object is not constructed at that point. This is
because C++ constructors are “real” constructors with regard to memory
population, and not mere initialisation callback methods as is the
case with for instance Ruby.

Therefore, each cActor subclass gets its own, private Init() method
with no inheritance. The constructors rely entirely on the regular
constructor call order.