Rework ground detection system.
In old OpenGL TSC, the checks for the ground object (the object an
object stands on that makes it immune to gravity) were done both on
falling and on moving, resulting in duplicate code.

The system introduced by this commit checks in
cActor::Handle_Collision_Massive() whether the object an actor collided
with is a valid ground object, and if so, places the actor on that
object, by calling Set_On_Ground(). These collisions are likely to be
caused by the gravity effect, implemented in cActor::Update_Gravity(),
that constantly tries to drag the actor downwards. Set_On_Ground() sets
mp_ground_object to the object the actor stands on, and places the actor
slightly above it (GROUND_HOVER_DISTANCE) so that not on each frame a
collision is generated between the standing actor and the ground
actor. GROUND_HOVER_DISTANCE is so small that the hovering effect is not
optically noticable (0.1f).

Once set on ground (detectable by whether mp_ground_object points to
something other than NULL), each time the actor moves it is checked
whether its collision rectangle if moved slightly downards (2 *
GROUND_HOVER_DISTANCE) still intersects the collision rectangle of the
bottom actor it is standing on. If it doesn’t, Check_On_Ground() will
call Reset_On_Ground() to release the connection, which makes the actor
subject to gravity again.

It might happen that the actor just passed from one ground actor to the
next (such as adjascent ground tiles), in which case the actor will be
subject to gravity for exactly one frame. However, because
Set_On_Ground() resets the actor’s position (via GROUND_HOVER_DISTANCE),
this does not pose a problem. A one-frame picture change should also not
be noticable.