Add update validity mechanism.
This is not complete yet. OpenGL TSC had a draw and an update validity
mechanism that were used to determine whether any given actor needed to
be updated and/or drawn. This was used to lower the amount of code
required to be executed per frame, most notably if actors were far out
of the camera range.