Add C++11 compilation support.
Newer Linux distros appear to compile some of our dependencies
in such a way that their headers include C++11-only statements.
Thus, if one tries to compile TSC (which does not need C++11
on itself) without C++11, it fails to compile on such systems,
because it includes headers which contain C++11 statements. Enabling
C++11 support with -std=c++11 resolves the problem.

-Wno-deprecated-declarations prevents massive amounts of warnings from
glibmmin C++11-mode that uses some features in its headers which are
deprecated in C++11.