Switch to bare sf::Window for OpenGL.
TSC currently interacts with the graphics card using raw OpenGL, not
using an intermediate library that wraps OpenGL calls (akin to SFML’s
Graphics module). This means that TSC’s use of SDL is restricted to
interaction with the user and the filesystem (and the music, but for
that see PR #488). Replacing those parts is probably easier than
directly imposing a new actor architecture as was tried previously in
the feature-sfml-port; most of the OpenGL code can be kept then.