Convert cVideo::Init_Image_Cache to SFML (mostly).
Replaces the use of SDL’s IMG_Load() during cache initialisation and
subsequent processing of SDL surfaces with sf::Image. Exempted from the
conversion is (for now marked with “// OLD”) printing of the filenames
being converted, because that requires a conversion of pFont to SFML
first, which is another big piece of work.

This is unfinished work. Do not try to compile.