Fix detection of invalid joystick.
Since with SFML joystick numbers are of type "unsigned int", -1 as the
default value has already been changed to 999 in an earlier commit (SFML
only supports up to 8 joysticks, so this is a sane indicator for an
invalid value). The condition check in this commit hadn’t been updated
to this, so this has now been fixed.