Fix bytes/bits per pixel/colour measure. Fixes #501.
SFML returns 8 bit per color in sf::Image::getPixelsPtr() according to
the docs, and an RGBA image. That is 8 bits x 4 colors (RGBA) = 32 bits
per pixel. 32 bits can be stored in 4 bytes (32/4). Thus the correct
value for image_bpp is 4.

The code in Downcase_Image() uses the SDL notation of *bytes* per *pixel*
rather than the more common notation of *bits* per *colour*. The
previous use of that method thought otherwise, and the result was a
segfault (wrong size of the image data array).

Confer:
  * https://wiki.libsdl.org/SDL_PixelFormat
  * https://en.wikipedia.org/wiki/Color_depth
  * http://www.sfml-dev.org/documentation/2.3.2/classsf_1_1Image.php#ac6137a608a9efaae2735c13ff259c214