Rewrite rendering of the points HUD element.
The HUD heavily relies on texts being regular TSC sprites one can Blit()
onto. With the switch to SFML this does no longer hold true. This commit
rewrites the HUD’s text architecture to cope with that and in parallel
fixes some problems it had with the secrecy principle. Most notably it
is not the parent’s task to submit the contents of its children to the
render queue. This has been moved into the children now, which are now
instances of the replacement class cHudText.

cMiniPointsText replaces the cPointsText class that was used to draw the
mini points (those appearing next to killed enemies). cStatusText was
made independant from the TSC sprite system so that it is *not* itself a
sprite anymore, but rather contains an sf::Text instance instead that it
submits to the render queue in its Draw() method. As a composed object
rather than being a sprite itself, it is also more justified to contain
children objects. A sprite containing other sprites which were drawn all
over the display as it was previously appears like incorrect code
architecture to me.