Port all the HUD elements to SFML.
The debug window has been disabled, because it would need a total
rewrite. As it should be rewritten to use CEGUI (or a future GUI system)
anyway that would mean a second rewrite. To prevent that, it is simply
removed for now. The code for the debug window can be retrieved from the
Git history at any time.

The HUD manager is not yet ported, thus this commit is not all needed
for porting the HUD.