Remove cMenu_Item class.
The menu handler now concerns itself only and only with managing the
amount of menu items and when they are selected. It has no involvement
in drawing the menu items anymore, nor does it say anything about what
to execute when a menu item is activated.

These tasks are now intended to be implemented by the respective menus
themselves, for which a new callback method
cMenu_Base::Selected_Item_Changed() was introduced for exactly that
purpose.